+++ reduction/tests/test_dead_time.py
@@ -1,6 +1,7 @@
 import mantid.simpleapi as mtd_api
+
 from lr_reduction import template
-from lr_reduction.DeadTimeCorrection import SingleReadoutDeadTimeCorrection
+from lr_reduction.dead_time_correction import SingleReadoutDeadTimeCorrection
 from lr_reduction.utils import amend_config
 
 mtd_api.config["default.facility"] = "SNS"
@@ -47,6 +48,33 @@ def test_deadtime_paralyzable(nexus_dir):
         assert c > 0
         assert c < 1.001
 
+def test_deadtime_threshold(nexus_dir):
+    """
+    Test using the threshold. Here the threshold is set to 0,
+    so all corrections will be 0
+    """
+    with amend_config(data_dir=nexus_dir):
+        ws = mtd_api.Load("REF_L_198409")
+
+    algo = SingleReadoutDeadTimeCorrection()
+    algo.PyInit()
+    algo.setProperty("InputWorkspace", ws)
+    algo.setProperty("OutputWorkspace", "dead_time_corr")
+
+    # set the deadtime threshold
+    # the mean corr is 1.0003, so use that as test value
+    algo.setProperty("UseDeadTimeThreshold", True)
+    algo.setProperty("DeadTimeThreshold", 1.0003)
+
+    algo.PyExec()
+    corr_ws = algo.getProperty("OutputWorkspace").value
+    corr = corr_ws.readY(0)
+
+    # manual inspection showed 88 corrections were above the threshold
+    assert len(corr[corr == 0]) == 88
+
+    for c in corr:
+        assert c <= 1.0003
 
 def test_full_reduction(nexus_dir):
     """